Version 0.2.22
Fixed for Windows 95: Retrieving environment variables should work now. The problem was caused by Win95's undocumented buffer size limit in GetEnvironmentVariable(). [Thanks to Robert Yaklin]

Fixed RWIN (the right windows key): When used with a hook-hotkey, it might have caused the Start Menu to appear.

Removed an unintended 10ms delay during the launch of every hotkey [introduced in v0.2.20].

Changed math operations (EnvAdd/Sub/Mult/Div) so that the output variable is made blank if either of the inputs is not purely numeric. In addition to making flawed scripts fail in a more consistent way, this behavior can be used to quickly determine whether or not a variable is purely numeric.

Improved RegRead and RegWrite to support REG_BINARY. [As with so many other things, thanks to Jon, the AutoIt author]

Improved the handling of ALT keys: if either of them is used as both a prefix (modifying other hotkeys) and a suffix (is itself a hotkey whose action fires when the key is released), it doesn't have the side-effect of invoking a menu on the active window. Strangely, this was never a problem on my system until today when some mysterious change occurred making it obvious this adjustment was needed.

Added support for floating point: Any number containing a decimal point is considered to be such. Any math operation involving at least one floating point number will produce a floating point result. In addition, the Random command has been improved to optionally generate floating point numbers. Note: integer math is still used whenever floating point numbers aren't involved because testing reveals that scripts with math intensive loops run over 5 times faster with 64-bit integers vs. floating point numbers (tested on Pentium II and Athlon XP).

Switched from 32 to 64-bit integer math. 32-bit is only about 40% faster than 64-bit for scripts running intensive math loops on Pentium IIs and Athlon XPs. But for real-world scripts, the difference is negligible, so the greater flexibility seems worth the cost. 64-bit also allows %A_TickCount% and DWORDs read from the registry to be correctly reported as positive, as well as enabling the retrieval of the exact number of bytes in files larger than 2 gigabytes.

Added command SetFormat to allow floating point numbers to be stored in a format other than the default (e.g. with blank padding, 0-padding, or a different number of decimal places).

Added Explorer rightclick >>New >>AutoHotkey Script.

+++ Source/window.cpp
@@ -978,7 +978,7 @@ int MsgBox(int aValue)
 
 
 
-int MsgBox(char *aText, UINT uType, char *aTitle, UINT aTimeout)
+int MsgBox(char *aText, UINT uType, char *aTitle, double aTimeout)
 // Returns FAIL if the attempt failed because of too many existing MessageBox windows,
 // or if MessageBox() itself failed.
 {
